+++ src/pages/edit.tsx
@@ -1,9 +1,18 @@
 import React from 'react'
+import AppBar from '@material-ui/core/AppBar'
+import Toolbar from '@material-ui/core/Toolbar'
+import Typography from '@material-ui/core/Typography'
+import InputBookForm from '../components/InputBookForm'
 
 export default function Edit() {
   return (
     <div>
-      <p>edit</p>
+      <AppBar position="static">
+        <Toolbar>
+          <Typography variant="h6">書籍編集</Typography>
+        </Toolbar>
+      </AppBar>
+      <InputBookForm />
     </div>
   )
 }
